+++ CHANGELOG.md
@@ -46,6 +46,7 @@ We haven't had the best track record of code/feature changes before this date, b
   - [fix] Avoid generating unnecessary empty strings in fbt result contents
   - [chore] Log type of leaf reached during table access failure
   - [fix] Fix FlowFixMes in FbtTable.access
+  - [feat] Improve feature detection of `Object.defineProperty()` for IE11
 </details>
 
 - 0.14.0: